FILE: renames data/ to game/
