+++ libs/spirit-islander/util/src/lib/functions/create-game-setup.ts
@@ -1,7 +1,7 @@
-import { SPIRITS } from '../data/spirits';
+import { SPIRITS } from '../game/spirits';
 import type { Config } from '../types/config.interface';
 import type { GameSetup } from '../types/game-setup.interface';
-import type { Difficulty } from '../data/difficulty';
+import type { Difficulty } from '../game/difficulty';
 import { getValidCombos } from './get-valid-combos';
 import { getOptionsByName } from './get-options';
 import { getDifficulty } from './get-difficulty';
