+++ libs/spirit-islander/util/src/lib/functions/get-adversary-by-id.ts
@@ -2,7 +2,7 @@ import {
   ADVERSARIES,
   AdversaryLevelId,
   AdversaryName,
-} from '../data/adversaries';
+} from '../game/adversaries';
 
 const adversaryLevelIdDict = ADVERSARIES.reduce(
   (accum, adversary) => {
