+++ libs/spirit-islander/util/src/lib/functions/get-difficulty.ts
@@ -1,5 +1,5 @@
-import type { Difficulty } from '../data/difficulty';
-import type { ExpansionName } from '../data/expansions';
+import type { Difficulty } from '../game/difficulty';
+import type { ExpansionName } from '../game/expansions';
 
 export function getDifficulty(
   difficulty: Difficulty | ((expansions: ExpansionName[]) => Difficulty),
