+++ libs/spirit-islander/util/src/lib/functions/update-model.ts
@@ -1,4 +1,4 @@
-import type { ExpansionName } from '../data/expansions';
+import type { ExpansionName } from '../game/expansions';
 
 export function updateModel<TName>(
   createModel: (expansions?: ExpansionName[]) => TName[],
