+++ libs/spirit-islander/util/src/lib/game/adversaries.ts
