+++ libs/spirit-islander/util/src/lib/game/boards.ts
