+++ libs/spirit-islander/util/src/lib/game/difficulty.ts
