+++ libs/spirit-islander/util/src/lib/game/maps.ts
