+++ libs/spirit-islander/util/src/lib/game/option.ts
