+++ libs/spirit-islander/util/src/lib/game/scenarios.ts
