+++ libs/spirit-islander/util/src/lib/game/spirits.ts
